# apps/web/src/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PrivacyDeleteResponse": {
        "properties": {
          "deleted": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Deleted",
            "type": "object"
          }
        },
        "required": [
          "deleted"
        ],
        "title": "PrivacyDeleteResponse",
        "type": "object"
      },
      "PrivacyExportResponse": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "data"
        ],
        "title": "PrivacyExportResponse",
        "type": "object"
      },
...
      "QueueAddRequest": {
        "description": "POST /api/queue body \u2014 accepts one or many URLs at once.\n\nAt least one of the two fields must be non-empty; the endpoint\nrejects empty payloads with 400 so the frontend can keep the\npaste-box input trivial.",
        "properties": {
          "job_url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Url"
          },
          "job_urls": {
            "anyOf": [
              {
                "items": {
                  "format": "uri",
                  "maxLength": 2083,
                  "minLength": 1,
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Urls"
          }
        },
        "title": "QueueAddRequest",
        "type": "object"
      },
      "QueueItem": {
        "description": "Row in GET /api/queue. Wraps askpicky.schemas.QueuedJob for\nthe HTTP layer so the response shape stays stable if the domain\nmodel grows fields we don't want to expose.",
        "properties": {
          "added_at": {
            "format": "date-time",
            "title": "Added At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "job_url": {
            "title": "Job Url",
            "type": "string"
          },
          "processed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed At"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "status": {
            "enum": [
              "pending",
              "processing",
              "done",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "job_url",
          "status",
          "added_at"
        ],
        "title": "QueueItem",
        "type": "object"
      },
      "QueueListResponse": {
        "properties": {
          "done_count": {
            "default": 0,
            "title": "Done Count",
            "type": "integer"
          },
          "failed_count": {
            "default": 0,
            "title": "Failed Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/QueueItem"
            },
            "title": "Items",
            "type": "array"
          },
          "pending_count": {
            "default": 0,
            "title": "Pending Count",
            "type": "integer"
          },
          "processing_count": {
            "default": 0,
            "title": "Processing Count",
            "type": "integer"
          }
        },
        "title": "QueueListResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Web surface for the AskPicky job-search assistant.",
    "title": "AskPicky API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/privacy/applications/{session_id}": {
      "delete": {
        "operationId": "delete_application_api_privacy_applications__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Application"
      }
    },
...
    "/api/privacy/export": {
      "get": {
        "operationId": "export_privacy_data_api_privacy_export_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_raw",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Raw",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Privacy Data"
      }
    },
...
    "/api/queue": {
      "get": {
        "operationId": "list_queue_api_queue_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Queue"
      },
      "post": {
        "operationId": "add_to_queue_api_queue_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueAddRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QueueItem"
                  },
                  "title": "Response Add To Queue Api Queue Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add To Queue"
      }
    },
...
    "/api/queue/{job_id}": {
      "delete": {
        "operationId": "remove_from_queue_api_queue__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove From Queue"
      }
    },
...
  }
}